Refactor DeckTab to encapsulate restock logic into DiscardUI
This moves the `RestockTask` nested class into its own file and
shifts the restock animation processing logic out of `DeckTab` into
`DiscardUI`, cleaning up `DeckTab`'s scope and rendering loop.
